feat(config): update environment configuration for development
- Change APP_ENV from local to development
- Update redis client from phpredis to predis
- Update redis host and port configuration
- Remove unused APP_PORT variable
- Add new compose.development.yaml for docker setup

This commit introduces a new docker compose file for development environment and updates the .env.development configuration to use predis as redis client and adjust the redis host and port. It also removes the unused APP_PORT variable and changes the APP_ENV value. The Dockerfile is updated to improve caching of composer dependencies.